The main window now built with window class objects
